Refresh Token Logic implementation

+++ app/src/main/java/com/example/nimble/ui/login/AuthViewModel.kt
@@ -2,22 +2,26 @@ package com.example.nimble.ui.login
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.nimble.database.repositories.AuthDatabaseRepository
 import com.example.nimble.datastore.AuthStorageRepository
 import com.example.nimble.model.RequestState
 import com.example.nimble.model.UserToken
+import com.example.nimble.network.SessionManager
 import com.example.nimble.network.repositories.AuthRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.flowOn
+import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
 class AuthViewModel(
     private val authRepository: AuthRepository,
-    private val authStorageRepository: AuthStorageRepository
+    private val authStorageRepository: AuthStorageRepository,
+    private val authDatabaseRepository: AuthDatabaseRepository,
 ) : ViewModel() {
 
     private val _emailValue = MutableStateFlow("")
@@ -29,9 +33,6 @@ class AuthViewModel(
     private val _loginRequestState = MutableStateFlow<RequestState?>(null)
     val loginRequestState = _loginRequestState.asStateFlow()
 
-    private val _userToken = MutableStateFlow<UserToken?>(null)
-    val userToken = _userToken.asStateFlow()
-
     private val _accessToken = MutableStateFlow<String?>(null)
     val accessToken = _accessToken.asStateFlow()
 
@@ -41,26 +42,18 @@ class AuthViewModel(
 
     private fun getUserAccessToken() {
         viewModelScope.launch {
-            authStorageRepository.getAccessToken().collect {
-                if (it.isNullOrEmpty().not()) {
-                    _accessToken.value = it
-                    getUserToken()
+            authStorageRepository.getAccessToken()
+                .flowOn(Dispatchers.IO)
+                .collect {
+                    if (it.isNullOrEmpty().not()) {
+                        authDatabaseRepository.getUserToken(it!!)
+                            .flowOn(Dispatchers.IO)
+                            .collect { userTokenData ->
+                                SessionManager.setUserToken(userTokenData)
+                                _accessToken.value = it
+                            }
+                    }
                 }
-            }
-        }
-    }
-
-    // TODO Implement correctly this method
-    private fun getUserToken() {
-        if (_accessToken.value.isNullOrEmpty().not()){
-            _userToken.value = UserToken(
-                id = "123456",
-                accessToken = _accessToken.value!!,
-                tokenType = "Bearer",
-                expiresIn = 3600,
-                refreshToken = "abcdefgh",
-                createdAt = System.currentTimeMillis()
-            )
         }
     }
 
@@ -82,16 +75,18 @@ class AuthViewModel(
                     if (response.isSuccessful) {
                         response.body()?.data?.let {
                             _loginRequestState.value = RequestState.Success
-                            _userToken.value = UserToken(
+                            _accessToken.value = it.attributes.accessToken
+                            SessionManager.setUserToken(UserToken(
                                 id = it.id,
                                 accessToken = it.attributes.accessToken,
                                 tokenType = it.attributes.tokenType,
                                 expiresIn = it.attributes.expiresIn,
                                 refreshToken = it.attributes.refreshToken,
                                 createdAt = it.attributes.createdAt
-                            )
+                            ))
                             withContext(Dispatchers.IO) {
                                 authStorageRepository.saveUserAccessToken(it.attributes.accessToken)
+                                authDatabaseRepository.saveUserToken(SessionManager.getUserToken()!!)
                             }
                         }
                     } else {
@@ -106,15 +101,27 @@ class AuthViewModel(
     }
 
     fun logout() {
-        viewModelScope.launch {
-            authRepository.logout(_userToken.value!!.accessToken)
-            authStorageRepository.deleteUserAccessToken()
-            clearVariables()
+        SessionManager.getUserToken()?.let { userToken ->
+            viewModelScope.launch {
+                authRepository.logout(userToken.accessToken)
+                    .onEach {
+                        if (it.isSuccessful) {
+                            authStorageRepository.deleteUserAccessToken()
+                            authDatabaseRepository.deleteUserToken(userToken)
+                        }
+                    }
+                    .flowOn(Dispatchers.IO)
+                    .collect {
+                        if (it.isSuccessful) {
+                            clearVariables()
+                        }
+                    }
+            }
         }
     }
 
     private fun clearVariables() {
-        _userToken.value = null
+        SessionManager.setUserToken(null)
         _accessToken.value = null
         _loginRequestState.value = null
         _emailValue.value = ""
